Project3/requirements.txt
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
dnspython==2.2.1
Flask==2.2.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
lightgbm==3.3.0
MarkupSafe==2.1.1
numpy==1.23.2
pandas==1.4.3
PyJWT==2.4.0
pymongo==4.2.0
python-dateutil==2.8.2
pytz==2022.2.1
pyupbit==0.2.33
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.1
six==1.16.0
threadpoolctl==3.1.0
urllib3==1.26.12
websockets==10.3
Werkzeug==2.2.2
wincertstore==0.2
zipp==3.8.1
